chore: separate Table sub components in multiple files to prevent fast refresh issues
