+++ .eslintignore
@@ -5,3 +5,4 @@
 /node_modules
 /playwright/.cache
 /playwright-report
+/storybook-static
