+++ src/components/table/TableHeader.tsx
@@ -1,0 +1,12 @@
+import type { TableProps } from './Table';
+import { TableRow } from './TableRow';
+
+export const TableHeader = ({ children, hasBorder, style }: TableProps) => {
+  return (
+    <thead>
+      <TableRow hasBorder={hasBorder} style={style}>
+        {children}
+      </TableRow>
+    </thead>
+  );
+};
