+++ src/components/table/tableContext.ts
@@ -1,0 +1,7 @@
+import { createContext, useContext } from 'react';
+
+export const TableContext = createContext({ hasBorder: true, color: '' });
+export function useTableContext() {
+  const context = useContext(TableContext);
+  return context;
+}
